requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
alembic==1.13.1
python-dotenv==1.0.1
click==8.1.7
Jinja2==3.1.3
PyYAML==6.0.1
black==24.2.0
pytest==8.0.2
pytest-cov==4.1.0
